+++ hypervisor/Makefile
@@ -132,6 +132,7 @@ C_SRCS += boot/reloc.c
 C_SRCS += arch/x86/init.c
 
 # configuration component
+C_SRCS += arch/x86/configs/vm_config.c
 ifeq ($(CONFIG_SHARING_MODE),y)
 C_SRCS += arch/x86/configs/sharing_config.c
 else ifeq ($(CONFIG_PARTITION_MODE),y)
